Moved some helper functions to  the test base class
