+++ worlds/oot_soh/test/bases.py
@@ -1,5 +1,7 @@
 from test.bases import WorldTestBase
-from ..Items import Items
+from ..Enums import Items, Regions
+from .. Items import SohItem
+from BaseClasses import ItemClassification as IC
 
 class SohTestBase(WorldTestBase):
     game = "Ship of Harkinian"
@@ -12,3 +14,33 @@ class SohTestBase(WorldTestBase):
         Also automatically award the item for convenience sake.
         """
         self.collect(self.world.create_item(Items.GLITCHED))
+
+    def get_bundle(self) -> tuple:
+        """
+        Get a state, region and world bundle using ROOT as region.
+
+        This will ignore age requirements on items because both ages can reach root.
+        """
+        return self.multiworld.state, Regions.ROOT, self.world
+    
+    def get_reg_bundle(self, region) -> tuple:
+        """
+        Get a state, region and world bundle using a specified region.
+
+        This can be used to enfore age requirements by setting CHILD_SPAWN or ADULT_SPAWN as the required region
+        """
+        return self.multiworld.state, region, self.world
+    
+    def create_item(self, item) -> SohItem:
+        """
+        Create a SohItem by name to collect without having to have it shuffled first
+        """
+        return SohItem(item, IC.progression, None, self.world.player)
+    
+    def sweep(self) -> None:
+        """
+        Sweep multiworld state
+
+        This is only necisery if you're running an assertion before collecting items
+        """
+        self.multiworld.state.sweep_for_advancements()
